+++ client/src/components/Cover/PendingCoverage/PendingCoverage.css
@@ -1,3 +1,67 @@
+.topWordssss {
+  display: flex;
+  justify-content: space-around;
+  width: 80%;
+  margin: auto;
+  padding-top: 2%;
+  padding-bottom: 2%;
+  font-family: "Open Sans", sans-serif;
+  font-style: normal;
+  font-weight: bold;
+  font-size: 10px;
+  color: #9e9e9e;
+}
+.middleWordssss {
+  display: flex;
+  width: 80%;
+  margin: auto;
+  background: white;
+  border-radius: 15px;
+  font-family: "Open Sans", sans-serif;
+  font-style: normal;
+  font-weight: normal;
+  font-size: 10px;
+  color: #9e9e9e;
+  margin-bottom: 4%;
+}
+
+.msame01 {
+  margin-left: 2%;
+}
+
+.msame00 {
+  margin-left: 23%;
+}
+
+.msame02 {
+  margin-left: 14%;
+}
+
+.tspecial {
+  color: black;
+  text-decoration: underline;
+  text-decoration-color: #3f3cc8;
+}
+
+a:link {
+  text-decoration: none;
+}
+
+a:visited {
+  text-decoration: none;
+}
+
+a:hover {
+  text-decoration: underline;
+}
+
+a:active {
+  text-decoration: underline;
+}
+a {
+  color: inherit;
+  text-decoration: none;
+}
 .shift1 {
   background-color: #ffffff;
   width: 76.5%;
