+++ client/src/components/Cover/PendingCoverage/PendingCoverage.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "./PendingCov.css";
+import "./PendingCoverage.css";
 
 const database = [
   {
@@ -19,7 +19,6 @@ const database = [
     message: "Going home for the weekend"
   }
 ];
-
 function processData(database) {
   const listItems = database.map((entry, index) => (
     <PendingCov
@@ -75,48 +74,45 @@ class PendingCov extends React.Component {
       </div>
     );
   }
-  // export default class PendingCov(props) {
-  //   return (
-  //     <div className="shift1">
-  //       <div className="time_loc1">
-  //         <div className="firstrow1">
-  //           <div className="bold_desk1">
-  //             <p className="desk">{props.desk}</p>
-  //           </div>
-  //           <div className="colorful_box1">
-  //             <p className="loc">{props.loc}</p>
-  //           </div>
-  //         </div>
-  //         <p className="date1">{props.date}</p>
-  //         <p className="time">{props.time}</p>
-  //       </div>
-  //       <div className="need_cov1">
-  //         <p className="needname">{props.needname}</p>
-  //       </div>
-  //       <div className="notes1">
-  //         <p className="message">{props.message}</p>
-  //       </div>
-  //     </div>
-  //   );
-  // };
-
-  // loginClick(e) {
-  //   console.log("In click function");
-  //   var approve = document.getElementById("approve");
-  //   var deny = document.getElementById("deny");
-  //   fetch("/login", {
-  //     method: "POST",
-  //     headers: {
-  //       Accept: "application/json",
-  //       "Content-Type": "application/json"
-  //     },
-  //     body: JSON.stringify({ approve: approve, deny: deny })
-  //   })
-  //     .then(response => {
-  //       return response.json();
-  //     })
-  //     .then(jsonResponse => {
-  //       console.log(jsonResponse);
-  //     });
 }
-export default PendingCov;
+class Cover extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      items: [
+        {
+          desk: "Front Desk",
+          loc: "Moffitt",
+          date: "Wednesday, March 6, 2020",
+          time: "3:00 PM - 5:00 PM",
+          needname: "Broco Lee",
+          message: "Going home for the weekend"
+        }
+      ]
+    };
+  }
+
+  render() {
+    return (
+      <div className="all">
+        <div className="topWordssss">
+          <h1 className="tspecial">Pending Coverage</h1>
+          <h1 className="tsame">
+            <a href="/cover/pendingsupervisor">Pending Supervisor Approval</a>
+          </h1>
+          <h1 className="tsame">
+            <a href="/cover/requesthistory">Request History</a>
+          </h1>
+        </div>
+        <div className="middleWordssss">
+          <h2 className="msame01">Time and Location</h2>
+          <h2 className="msame00">Needing Coverage</h2>
+          <h2 className="msame02">Notes</h2>
+          <h2 className="msame03"></h2>
+        </div>
+        {processData(this.state.items)}
+      </div>
+    );
+  }
+}
+export default Cover;
